+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 * New option to show module's status in logs & other minor improvements
 * JSONPath option to extract temperature from JSON payload for MQTT Current Temperature Subscription
 * Show module's fault status in the log and publish to MQTT
+* Pushover notifications for temperature exceeding thresholds, connection errors, faults & incorrrect day/time
 
 ## 3.1.2 (2020-10-20)